backend/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyMySQL==1.1.2
SQLAlchemy==2.0.43
typing_extensions==4.15.0
python-dotenv==1.0.1
requests==2.32.3
PyJWT==2.9.0
PyMuPDF==1.24.0